+++ frontend/src/App.jsx
@@ -1,12 +1,13 @@
 import React from 'react';
-import AppRoutes from './routes/index.js'; // ou './routes/index' se você nomeou o arquivo como index.jsx
+//import AppRoutes from './routes/index.jsx'; // ou './routes/index' se você nomeou o arquivo como index.jsx
+import LandingPage from './pages/LandingPage.jsx'
 
-import './styles/App.css';
+
 
 function App() {
   return (
     <div className="app">
-      <AppRoutes />
+      <LandingPage />
     </div>
   );
 }
